Support for friend invite URLs, game draws.
A link with a game ID parameter now automatically tries to enter private
game. The invite a friend page now works and clicking on the email
button opens an email program with a link to the game in the body.

Game draws were failing and are fixed now, with some changes to
the related notifications (no_moves becomes draw).

Also consolidating handling of seek notifications a bit, now including
all responses to seek related commands.